reverse-ws/pyproject.toml
[tool.poetry]
name = "pixiv-bot"
version = "1.1.0a2"
description = "pixiv-bot"
authors = "ssttkkl <huang.wen.long@hotmail.com>"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0b4"
nonebot-plugin-pixivbot-onebot-v11 = "^1.1.0a2"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.7"

[tool.nonebot]
plugins = ["nonebot_plugin_pixivbot_onebot_v11"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
